完成持久化登录

+++ src/main/java/com/practice/community/community/controller/PublishController.java
@@ -74,7 +74,7 @@ public class PublishController {
         question.setDescription(description);
         question.setTag(tag);
         question.setCreator(user.getId());
-        question.setGmtCreate((int) System.currentTimeMillis());
+        question.setGmtCreate(System.currentTimeMillis());
         question.setGmtModified(question.getGmtCreate());
         questionMapper.insert(question);
         return "redirect:/";
