+++ src/main/java/com/practice/community/community/model/Question.java
@@ -12,8 +12,8 @@ public class Question {
     private Integer id;
     private String title;
     private String description;
-    private Integer gmtCreate;
-    private Integer gmtModified;
+    private Long gmtCreate;
+    private Long gmtModified;
     private Integer creator;
     private Integer commentCount;
     private Integer viewCount;
